third_party/WebKit/LayoutTests/external/wpt/css/cssom/style-sheet-interfaces-001-expected.txt

This is a testharness.js-based test.
PASS sheet_property
PASS sheet_property_updates
PASS CSSStyleSheet_properties
PASS CSSStyleSheet_property_values
PASS StyleSheet_properties
FAIL StyleSheet_property_values assert_equals: expected "all" but got "alla"
Harness: the test ran to completion.
